fix errors

+++ src/rational.cpp
@@ -112,16 +112,18 @@ namespace cyy::math {
     auto prime_c = primes();
 
     for (auto prime : prime_c.till(100)) {
-      while (p % prime == 0 && q % prime == 0) {
-        p /= prime;
-        q /= prime;
+      while (p % static_cast<uint32_t>(prime) == 0 && q % static_cast<uint32_t>(prime)== 0) {
+        p /=static_cast<uint32_t>(prime);
+        q /=static_cast<uint32_t>(prime);
       }
     }
     return *this;
+    /*
     auto gcd_int = gcd(p, q);
     p /= gcd_int;
     q /= gcd_int;
     return *this;
+    */
   }
 
   std::string rational::to_string() const {
